Removed the ignore option to exclude directories named `_common`.
This was for when test common files were placed under a directory which other test files resided. In order to distinguish between component class files and test files, the exclude option was set. But this is no longer needed.

For cases that need to set exclusion directories and other options, now the third parameter accepts options.