[Internal] Change the plugin site temporary directory path.
With the current temporary directory path, it caused permission conflicts; on a shared server, when one user creates with the plugin temporary base directory, WPAAL, with the permission 0755 restricted with umask(), another user cannot create a plugin site directory under it. So it's better to create own one directly under the system temporary directory.