Optimize the lock mechanism.
Currently, if the current time reaches the set timeout time, the file is considered unlocked. Change it to that "if the current time exceeds the set timeout time".

This is because it is confirmed that in some occasions, PA-API requests are simultaneously performed at the exact time within one second interval, which must be avoided.