Feature: Have form fields class the ability to access the factory class
So that the saved form option values can be retrieved. This is important because the selected locale can be known and thus whether the PA-API keys are set as well in order to show a warning to the user about PA-API requirements.

This is an additional change to fb0bdb57.