Feature: Add the ability to suppress requests for certain type of URLs
With the current design, Web Page Dumper is only used when an error is responded. However, there are cases that responses return the 200 status code (success) with an empty body. In that case, it is not detected as an error. So it is easier and less resource consuming to cancel the original request and perform a substitute request with Web Page Dumper.