Fix: Fix a bug with the White and Black List options
Combinations of White List ASIN/Title/Description and Black List ASIN/Title/Description options do not provide expected results. For example, set * in the Black List Title option and some keywords to the White List Description option. It does not keep the products that matches the description keywords.

This change should fix it.