Chore: Exclude the test component from exporting
Dur to the wordpress.org plugin review guideline, quoted below.

---
## Included Unneeded Folders

Your plugin includes folders and files that are not required for the running of your plugin. Some examples are:

    vendor folders (bower, node, grunt, etc)
    documentation
    demos
    unit tests

The entire library isn't going to be used by your users, and having all those files included is just extra weight to a plugin.

If you're trying to include the human-readable version of your own code (in order to comply with our guidelines) remember that we permit you to put links to them in your readme.

https://developer.wordpress.org/plugins/wordpress-org/detailed-plugin-guidelines/#4-code-must-be-mostly-human-readable

If you’ve made a Block plugin and used npm and webpack to compress and minify it, you can either include the source code within the published plugin or provide access to a public maintained source that can be reviewed, studied, and yes, forked.

But you can, and should, safely remove those other folders from your plugins.

Example(s) from your plugin:

amazon-auto-links/include/core/component/test/