Release: Rename the plugin name from Amazon Auto Links to Auto Amazon Links
Due to the complainment from wordpress.org plugin review team quoted below.

---
## Display Name Infringes on Trademarks

Your plugin's display name is infringing on someone's trademark or commonly known brand.

amazon-auto-links/readme.txt:1:=== Amazon Auto Links - Amazon Associates Affiliate Plugin ===
amazon-auto-links/amazon-auto-links.php:3: * Plugin Name: Amazon Auto Links

This can be a little long, so please read carefully. Most of your questions are answered.

The tl;dr is this:

    Your plugin display name may not begin with nor use the following term(s) in a manner that is confusing or implies a relationship: Amazon
    We recommend you use the following Display Name instead: Auto Links for Amazon Associates

Trademarks apply to the following aspects of your plugin:

    The Display Name - You may not begin the display name with someone else's trademarked (or commonly recognized) term
    All Images - You may not use trademarked logos/images in your banner, screenshots, logos, etc
    Permalinks - URLs may not contain or misuse trademarked terms

In order to comply with trademark law, we need you to update the readme file and the main plugin file to no longer infringe on their trademark. If your banners or icons include official logos, now is the time to clean that up too.

We know this can be confusing. In many cases it’s acceptable to use the ’shortname’ of a project or company for the plugin permalink (i.e. “ga” for “Google Analytics” or “ms” for “Microsoft”). However even if you can use it in your permalink, you MAY NOT use that short term to begin the display name.