Fix: Sanitize form inputs before saving
Since the implementation of the `Security` setting section, allowed HTML tags needs to be referred to a different option element. This change fixes the issue that some methods still reading the old Form option values.

Also, there is a line that skips formatting when an array element is not set and it was checking non-existent variable so it's possible that the formatting wasn't applied at all.