Optimize: Remove the `required` attribute from contextual widget setting fields
The `required` attribute tells the browser to show a pop-up notice to the user when the subject input does not have a value. However, for some reasons, in widget fields, this behavior seems to be disabled. So the user does not notice anything but the form cannot be submitted, which may make the user think it is broken.