Fix: Fix a bug of incorrect discount percentages
This bug caused by a value of `0`, retrieved from the table column `discounted_price`, whose type is `bigint(200)`. The integer column converts values to an integer when an empty string is set. So when the plugin set an empty string, it always sets `0`.

Currently, the plugin considers it as the real 0 price and causes the wrong discount rate. To prevent that, check the value of the `discounted_price_formatted` as well.