Fix: Fix a bug that some dynamic options which should be hidden are not hidden
This seems to be due to recent update of the Revealer custom field type. The script for the Revealer field type must load earlier but somehow, it loads late, which seems to cause the problem.

For a first aid, a delayed function call is applied. But it's not the cleanest solution. So it needs to find a better approach.