Feature: Add the `image` button type
From this change, the mechanism of button types are added. And the existing buttons are categorized as either `classic` or `theme`. This change adds the `image` type.

There are still unfinished part which includes:
 - the button preview in unit editing page (this requires overhaul of the existing script)
 - the ability to generate default buttons

It is planned to add the `button2` button type which can add SVG icons. Also, the `theme` button type needs to be enhanced to support SVG icons.