Fix: Send the body height to the parent frame rather than the button height
This is because, when a theme sets a CSS `line-height` property to the `body` tag, the body height are different from the button height, which causes a cut-off of the button in the preview frame. So frame height should be adjusted to the body height.

On the other hand, the width should constrain to the button width as the frame width can be larger than the button width.