Optimize: Check width changes of button preview as well with the `ReloadButtonPreview` custom event
Currently, it sends a window message from the frame window only when the height becomes different. This commit includes a change that also checks the width and if the width and height are both the same, it cancels sending the message.

The original reason of this change is that in the button preview script of the parent frame, when the `Override the button label` unit option is enabled and the user types something, the preview button's width changes. Thus, the iframe proportion needs to be adjusted accordingly. However, communications between frames through window messages are somewhat slow so the method of updating the frame dimensions has been switched directly querying the child frame content window DOM structure. So this commit change does not serve anything but just for a future possible implementation that involves with width changes of button previews.