Fix: Fix an issue that button previews in the list table for trashed ones do not load their styles
This is because the active button CSS rules are cached and those rules are loaded but the cache does not include trashed buttons.

This change adds the `load-own-style` `$_GET` URL query key for trashed buttons and if the query is given, it loads the CSS rules stored in the meta to the head tag of the framed page.