Feature: Introduce the new `aal_filter_output` and `aal_action_output` filters
These are a replacement for the `AmazonAutoLinks()` function which can be called by third-party developers. However, it should not be recommended as it causes an error of "Function not found" if the plugin is deactivated and the `if (function_exists( 'AmazonAutoLinks' ) )` is not used.

With these new hooks, developers can simply call, `do_action( 'aal_action_output', arguments );` or `echo apply_filters( 'aal_filter_output' );`. And they don't have to place `function_exists()` checks nor worry about errors after deactivating the plugin.