Optimize: Change the behavior of the getUnlockTime() method not to create a lock file if not exist
The purpose of this method is to retrieve the unlock-time. So it's not intuitive that it creates a file if it does not exist. With this change just `0` will be returned if the lock file does not exist. If a lock file must exist, call the `lock()` method in prior.