Feature: Change the timing of applying the sort order of PA-API Item-Lookup units
This way, when setting the `random` sort order, the random order is applied to all the specified items. Before this change, only it was applied to fetched items with the `count` argument, meaning that when the count is 2 and 3 products are fetched, the random order was applied only to the first two. Now with this change the randomness is applied to all the three first. Then, two will be returned.